Improvements

+++ longest_seqs.cpp
@@ -13,15 +13,15 @@ a given sequence such that all elements of the subsequence are sorted in increas
 
 For example, length of LIS for {10, 22, 9, 33, 21, 50, 41, 60, 80} is 6 and LIS is {10, 22, 33, 50, 60, 80}.
  */
-uint32_t LISLen(const int32_t a[], const uint32_t n) {
+uint LISLen(const int a[], const uint n) {
   if (!n)
     return 0;
 
-  vector<uint32_t> dp(n, 1); // initialize to 1, as each number is a sequence
-  uint32_t ret = 1;
+  vector<uint> dp(n, 1); // initialize to 1, as each number is a sequence
+  uint ret = 1;
 
-  for (uint32_t i = 1; i < n; ++i) {
-    for (uint32_t j = 0; j < i; ++j) {
+  for (uint i = 1; i < n; ++i) {
+    for (uint j = 0; j < i; ++j) {
       if (a[i] > a[j]) {
         dp[i] = max(dp[i], dp[j]+1);
       }
@@ -47,29 +47,29 @@ In this grid, (3, 2, 1, 0, 1) is a snake sequence.
 Given a grid, find the longest snake sequences and their lengths
 (so there can be multiple snake sequences with the maximum length).
  */
-uint32_t longestSnakeSeqLen(const vector<vector<int32_t>>& mat) {
-  const uint32_t m = mat.size();
-  const uint32_t n = mat[0].size();
+uint longestSnakeSeqLen(const vector<vector<int>>& mat) {
+  const uint m = mat.size();
+  const uint n = mat[0].size();
   if (m < 2 || n < 2)
     return 0;
 
-  uint32_t ret = 0;
-  vector<vector<uint32_t>> dp(m, vector<uint32_t>(n, 1));
+  uint ret = 0;
+  vector<vector<uint>> dp(m, vector<uint>(n, 1));
 
-  for (int32_t i = m-2; i >= 0; --i) { // right most column, can't move right
+  for (int i = m-2; i >= 0; --i) { // right most column, can't move right
     if (abs(mat[i][n-1]-mat[i+1][n-1]) == 1) {
       dp[i][n-1] = dp[i+1][n-1]+1;
       ret = max(dp[i][n-1], ret);
     }
   }
-  for (int32_t j = n-2; j >= 0; --j) { // bottom row, can't move down
+  for (int j = n-2; j >= 0; --j) { // bottom row, can't move down
     if (abs(mat[m-1][j]-mat[m-1][j+1]) == 1) {
       dp[m-1][j] = dp[m-1][j+1]+1;
       ret = max(dp[m-1][j], ret);
     }
   }
-  for (int32_t i = m-2; i >= 0; --i) { // now move
-    for (int32_t j = n-2; j >= 0; --j) {
+  for (int i = m-2; i >= 0; --i) { // now move
+    for (int j = n-2; j >= 0; --j) {
       if (abs(mat[i][j]-mat[i][j+1]) == 1) {
         dp[i][j] = dp[i][j+1]+1;
       }
@@ -96,22 +96,22 @@ Given
 
 The output should be [1, 2, 3, 4, 5, 6, 7, 8]
 
-vector<int32_t> matrixLIS(const vector<vector<int32_t>>& mat) {
-  const uint32_t m = mat.size();
-  const uint32_t n = mat[0].size();
+vector<int> matrixLIS(const vector<vector<int>>& mat) {
+  const uint m = mat.size();
+  const uint n = mat[0].size();
   if (!m || !n)
     return {};
 
-  static const vector<pair<int8_t,int8_t>> dirs = {{-1,0}, {1,0}, {0,-1}, {0,1}};
-  vector<vector<int32_t>> dp(m, vector<int32_t>(n, 0));
+  static const vector<pair<int,int>> dirs = {{-1,0}, {1,0}, {0,-1}, {0,1}};
+  vector<vector<int>> dp(m, vector<int>(n, 0));
 
-  function<int32_t(const uint32_t,const uint32_t)> solve = // DFS
-    [&](const uint32_t i, const uint32_t j) {
+  function<int(const uint,const uint)> solve = // DFS
+    [&](const uint i, const uint j) {
     if (dp[i][j] != 0)
       return dp[i][j];
 
     for (const auto& d : dirs) {
-      int32_t newi = i+d.first, newj = j+d.second;
+      int newi = i+d.first, newj = j+d.second;
       if (newi < 0 || newj < 0 || newi >= m || newj >= n)
         continue;
 
@@ -122,10 +122,10 @@ vector<int32_t> matrixLIS(const vector<vector<int32_t>>& mat) {
     return ++dp[i][j];
   };
 
-  int32_t max_start = 0, max_path = 0;
-  for (uint32_t i = 0; i < m; ++i) {
-    for (uint32_t j = 0; j < n; ++j) {
-      int32_t path = solve(i, j);
+  int max_start = 0, max_path = 0;
+  for (uint i = 0; i < m; ++i) {
+    for (uint j = 0; j < n; ++j) {
+      int path = solve(i, j);
       if (path > max_path) {
         max_start = mat[i][j];
         max_path = path;
@@ -133,7 +133,7 @@ vector<int32_t> matrixLIS(const vector<vector<int32_t>>& mat) {
     }
   }
 
-  vector<int32_t> ret(max_path);
+  vector<int> ret(max_path);
   iota(ret.begin(), ret.end(), max_start);
   return ret;
 }
@@ -155,13 +155,13 @@ Output: 5 (A Longest Bitonic Subsequence of length 5 is 12, 11, 5, 3, 1)
 Input a[] = {80, 60, 30, 40, 20, 10}
 Output: 5 (A Longest Bitonic Subsequence of length 5 is 80, 60, 30, 20, 10)
  */
-uint32_t longestBitonicSubseqLen(const uint32_t a[], const uint32_t n) {
+uint longestBitonicSubseqLen(const uint a[], const uint n) {
   if (n < 3)
     return 0;
 
-  vector<uint32_t> t1(n, 1), t2(n, 1);
-  for (int32_t i = 0, k = n-1; i < n && k >= 0; ++i, --k) {
-    for (int32_t j = i+1, l = k-1; j < n && l >= 0; ++j, --l) {
+  vector<uint> t1(n, 1), t2(n, 1);
+  for (int i = 0, k = n-1; i < n && k >= 0; ++i, --k) {
+    for (int j = i+1, l = k-1; j < n && l >= 0; ++j, --l) {
       if (a[j] > a[i])
         t1[j] = t1[i]+1;
       if (a[l] > a[k])
@@ -173,8 +173,8 @@ uint32_t longestBitonicSubseqLen(const uint32_t a[], const uint32_t n) {
       t2[k] = max(t2[k], t2[k+1]);
   }
 
-  uint32_t len = 0;
-  for (uint32_t i = 1; i < n; ++i) {
+  uint len = 0;
+  for (uint i = 1; i < n; ++i) {
     len = max(t1[i-1]+t2[i], len);
   }
 
@@ -182,23 +182,23 @@ uint32_t longestBitonicSubseqLen(const uint32_t a[], const uint32_t n) {
 }
 
 int main(int argc, char** argv) {
-  int32_t a[9] = {10, 22, 9, 33, 21, 50, 41, 60, 80};
+  int a[9] = {10, 22, 9, 33, 21, 50, 41, 60, 80};
   cout << "LIS length: " << LISLen(a, 9) << endl;
 
-  vector<vector<int32_t>> mat = {{1, 3, 2, 6, 8},
+  vector<vector<int>> mat = {{1, 3, 2, 6, 8},
                                  {-9, 7, 1, -1, 2},
                                  {1, 5, 0, 1, 9}};
   cout << "Longest snake sequence length: " << longestSnakeSeqLen(mat) << endl;
 
  /* mat = {{1,8}, {2,7}, {3,6}, {4,5}};
-    vector<int32_t> res = matrixLIS(mat);
+    vector<int> res = matrixLIS(mat);
 
   cout << "Longest increasing sequence: ";
   for (const auto i : res)
     cout << i << " ";
   cout << endl;
  */
-  uint32_t b[8] = {1, 11, 2, 10, 4, 5, 2, 1};
+  uint b[8] = {1, 11, 2, 10, 4, 5, 2, 1};
   cout << "Longest bitonic subsequence length is: " << longestBitonicSubseqLen(b, 8) << endl;
 
   return 0;
